# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface NetworkStatus {
  chainId: number;
  network: string;
  blockNumber: number;
  gasPrice: string;
  gasPriceGwei: string;
  isMainnet: boolean;
  rpcUrl: string;
  explorerUrl: string;
  status: string;
}

export interface TokenBalance {
  symbol: string;
  name: string;
  address: string;
  balance: string;
  decimals: number;
  usdValue?: string | null;
}

export interface WalletBalance {
  address: string;
  ethBalance: string;
  ethBalanceFormatted: string;
  usdValue?: string | null;
  tokens: TokenBalance[];
  network: string;
  explorerUrl: string;
}

export interface TransactionInfo {
  hash: string;
  blockNumber?: number | null;
  from: string;
  to?: string | null;
  value: string;
  valueFormatted: string;
  gasPrice: string;
  gasUsed?: string | null;
  status: string;
  timestamp?: number | null;
  explorerUrl: string;
  network: string;
}

export interface TokenInfo {
  symbol: string;
  name: string;
  address: string;
  price: string;
  priceChange24h: string;
  volume24h?: string | null;
  decimals: number;
  logoUrl?: string | null;
}

export interface TokenList {
  tokens: TokenInfo[];
  network: string;
  updatedAt: string;
}

export interface ProtocolInfo {
  name: string;
  category: string;
  description: string;
  website: string;
  tvl?: string | null;
  logoUrl?: string | null;
  contractAddress?: string | null;
  isActive: boolean;
}

export interface ProtocolList {
  protocols: ProtocolInfo[];
  network: string;
}

/**
 * Which Arbitrum network to register on
 */
export type RegisterAgentRequestNetwork =
  (typeof RegisterAgentRequestNetwork)[keyof typeof RegisterAgentRequestNetwork];

export const RegisterAgentRequestNetwork = {
  mainnet: "mainnet",
  sepolia: "sepolia",
} as const;

export interface RegisterAgentRequest {
  /** Agent name */
  name: string;
  /** Agent description */
  description: string;
  /** Private key of the wallet to use for registration (handled server-side, never stored) */
  privateKey: string;
  /** Which Arbitrum network to register on */
  network: RegisterAgentRequestNetwork;
  /** Optional URI pointing to agent metadata */
  metadataUri?: string | null;
}

export interface RegisterAgentResponse {
  success: boolean;
  transactionHash?: string | null;
  agentAddress: string;
  registryAddress: string;
  network: string;
  explorerUrl?: string | null;
  message: string;
}

export interface AgentStatus {
  address: string;
  isRegistered: boolean;
  registryAddress: string;
  network: string;
  registrationBlock?: number | null;
  explorerUrl: string;
}

export type McpToolInputSchema = { [key: string]: unknown };

export interface McpTool {
  name: string;
  description: string;
  inputSchema: McpToolInputSchema;
}

export interface McpToolList {
  tools: McpTool[];
  version: string;
  description: string;
}

/**
 * Tool arguments
 */
export type McpExecuteRequestArguments = { [key: string]: unknown };

export interface McpExecuteRequest {
  /** Tool name to execute */
  tool: string;
  /** Tool arguments */
  arguments: McpExecuteRequestArguments;
}

export type McpExecuteResponseResult = { [key: string]: unknown } | null;

export interface McpExecuteResponse {
  success: boolean;
  result?: McpExecuteResponseResult;
  error?: string | null;
  tool: string;
  executedAt: string;
}

export interface OverviewStats {
  network: string;
  blockNumber: number;
  totalToolCalls: number;
  activeAgents: number;
  supportedProtocols: number;
  availableTools: number;
  avgGasPrice: string;
  uptimePercent: string;
}