# lib/api-zod/src/generated/types/agentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentStatus {
  address: string;
  isRegistered: boolean;
  registryAddress: string;
  network: string;
  registrationBlock?: number | null;
  explorerUrl: string;
}