# lib/api-zod/src/generated/types/mcpExecuteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { McpExecuteRequestArguments } from "./mcpExecuteRequestArguments";

export interface McpExecuteRequest {
  /** Tool name to execute */
  tool: string;
  /** Tool arguments */
  arguments: McpExecuteRequestArguments;
}