# lib/api-zod/src/generated/types/mcpExecuteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { McpExecuteResponseResult } from "./mcpExecuteResponseResult";

export interface McpExecuteResponse {
  success: boolean;
  result?: McpExecuteResponseResult;
  error?: string | null;
  tool: string;
  executedAt: string;
}