# lib/api-zod/src/generated/types/mcpTool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { McpToolInputSchema } from "./mcpToolInputSchema";

export interface McpTool {
  name: string;
  description: string;
  inputSchema: McpToolInputSchema;
}