# lib/api-zod/src/generated/types/mcpToolList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { McpTool } from "./mcpTool";

export interface McpToolList {
  tools: McpTool[];
  version: string;
  description: string;
}