# lib/api-zod/src/generated/types/overviewStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */

export interface OverviewStats {
  network: string;
  blockNumber: number;
  totalToolCalls: number;
  activeAgents: number;
  supportedProtocols: number;
  availableTools: number;
  avgGasPrice: string;
  uptimePercent: string;
}