# lib/api-zod/src/generated/types/protocolList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { ProtocolInfo } from "./protocolInfo";

export interface ProtocolList {
  protocols: ProtocolInfo[];
  network: string;
}