# lib/api-zod/src/generated/types/registerAgentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterAgentRequestNetwork } from "./registerAgentRequestNetwork";

export interface RegisterAgentRequest {
  /** Agent name */
  name: string;
  /** Agent description */
  description: string;
  /** Private key of the wallet to use for registration (handled server-side, never stored) */
  privateKey: string;
  /** Which Arbitrum network to register on */
  network: RegisterAgentRequestNetwork;
  /** Optional URI pointing to agent metadata */
  metadataUri?: string | null;
}