# lib/api-zod/src/generated/types/registerAgentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterAgentResponse {
  success: boolean;
  transactionHash?: string | null;
  agentAddress: string;
  registryAddress: string;
  network: string;
  explorerUrl?: string | null;
  message: string;
}