# lib/api-zod/src/generated/types/tokenBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenBalance {
  symbol: string;
  name: string;
  address: string;
  balance: string;
  decimals: number;
  usdValue?: string | null;
}