# lib/api-zod/src/generated/types/tokenList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */
import type { TokenInfo } from "./tokenInfo";

export interface TokenList {
  tokens: TokenInfo[];
  network: string;
  updatedAt: string;
}