# lib/api-zod/src/generated/types/transactionInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArbiLink - Arbitrum MCP Server API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionInfo {
  hash: string;
  blockNumber?: number | null;
  from: string;
  to?: string | null;
  value: string;
  valueFormatted: string;
  gasPrice: string;
  gasUsed?: string | null;
  status: string;
  timestamp?: number | null;
  explorerUrl: string;
  network: string;
}